数据库脚步迁移

+++ modules/core/src/Heal.Core.Domain/Bases/Organizations/Models/OrganizationWithChildCount.cs
@@ -5,4 +5,4 @@
 /// </summary>
 /// <param name="Id">Id</param>
 /// <param name="Count">数量</param>
-public record OrganizationWithChildCount(Guid Id, int Count);
+public record OrganizationWithChildCount(Guid Id, long Count);
